backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL adapter
openpyxl==3.1.5  # Excel file generation
slowapi==0.1.9  # Rate limiting for API endpoints

